Update select to rmpv 0.3.0.

+++ src/lib.rs
@@ -60,7 +60,7 @@ pub use rmpv::{Utf8String, Integer};
 
 
 pub const FIX_STR_PREFIX: u8 = 0xA1;
-pub const TARANTOOL_SPACE_ID: u16 = 280;
-pub const TARANTOOL_SPACE_ID_KEY_NUMBER: u8 = 2;
-pub const TARANTOOL_INDEX_ID: u16 = 288;
-pub const TARANTOOL_INDEX_ID_KEY_NUMBER: u8 = 2;
+pub const TARANTOOL_SPACE_ID: u64 = 280;
+pub const TARANTOOL_SPACE_ID_KEY_NUMBER: u64 = 2;
+pub const TARANTOOL_INDEX_ID: u64 = 288;
+pub const TARANTOOL_INDEX_ID_KEY_NUMBER: u64 = 2;
